requirements.txt
# Core Agent Development Kit
google-genai>=1.0.0

# Data Analysis & Visualization
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Data Processing
openpyxl>=3.1.0  # Excel support
xlrd>=2.0.1  # Legacy Excel support

# Observability & Logging
structlog>=23.1.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.2.0

# Optional: Database support
sqlalchemy>=2.0.0
pymysql>=1.1.0
psycopg2-binary>=2.9.0

# Development
pytest>=7.4.0
black>=23.7.0

